beehive-datagram-loader/requirements.txt
psycopg2-binary
pika>=1.0.0
git+https://github.com/waggle-sensor/pywaggle@v0.25.0
